refactor: tests in src and utils in folder
