--- src/integration-tests/src/tests/mod.rs
@@ -1,9 +1,0 @@
-//! Tests are named with letters to keep them in alphabetical order and ensure they are executed in that order.
-//! Make sure you run tests with `--test-threads=1` so that they are executed sequentially.
-
-pub mod a_ws_open;
-pub mod b_ws_message;
-pub mod c_ws_get_messages;
-pub mod d_ws_close;
-pub mod e_ws_send;
-pub mod f_messages_acknowledgement;
