+++ src/integration-tests/src/utils/wasm.rs
